+++ src/ApiClient.php
@@ -9,12 +9,18 @@ namespace Wizaplace\SDK;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\ClientException;
+use GuzzleHttp\RequestOptions;
 use Jean85\PrettyVersions;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\UriInterface;
 use Wizaplace\SDK\Authentication\ApiKey;
 use Wizaplace\SDK\Authentication\AuthenticationRequired;
 use Wizaplace\SDK\Authentication\BadCredentials;
+use Wizaplace\SDK\Exception\BasketNotFound;
+use Wizaplace\SDK\Exception\CouponCodeAlreadyApplied;
+use Wizaplace\SDK\Exception\CouponCodeDoesNotApply;
+use Wizaplace\SDK\Exception\DomainError;
+use Wizaplace\SDK\Exception\ErrorCode;
 use Wizaplace\SDK\Exception\JsonDecodingError;
 
 final class ApiClient
@@ -27,6 +33,9 @@ final class ApiClient
 
     /** @var string */
     private $version;
+
+    /** @var null|string */
+    private $language;
 
     public function __construct(Client $client)
     {
@@ -126,14 +135,65 @@ final class ApiClient
      */
     public function rawRequest(string $method, $uri, array $options = []): ResponseInterface
     {
-        $options['headers']['User-Agent'] = 'Wizaplace-PHP-SDK/'.$this->version;
+        $options[RequestOptions::HEADERS]['User-Agent'] = 'Wizaplace-PHP-SDK/'.$this->version;
+        if ($this->language !== null) {
+            $options[RequestOptions::HEADERS]['Accept-Language'] = $this->language;
+        }
 
-        return $this->httpClient->request($method, $uri, $this->addAuth($options));
+        try {
+            return $this->httpClient->request($method, $uri, $this->addAuth($options));
+        } catch (ClientException $e) {
+            $domainError = $this->extractDomainErrorFromClientException($e);
+            if ($domainError !== null) {
+                throw $domainError;
+            }
+
+            throw $e;
+        }
     }
 
     public function getBaseUri(): ?UriInterface
     {
         return $this->httpClient->getConfig('base_uri');
+    }
+
+    /**
+     * Changes the language the responses' content will be in.
+     *
+     * @param null|string $language the language code, for example 'fr' or 'en'
+     */
+    public function setLanguage(?string $language): void
+    {
+        $this->language = $language;
+    }
+
+    private function extractDomainErrorFromClientException(ClientException $e): ?DomainError
+    {
+        try {
+            $response = $this->jsonDecode($e->getResponse()->getBody()->getContents(), true);
+
+            if (!isset($response['error'])) {
+                return null;
+            }
+
+            $errorCode = new ErrorCode($response['error']['code']);
+
+            $errorsClasses = [
+                BasketNotFound::class,
+                CouponCodeDoesNotApply::class,
+                CouponCodeAlreadyApplied::class,
+            ];
+
+            foreach ($errorsClasses as $errorClass) {
+                if ($errorClass::getErrorCode()->equals($errorCode)) {
+                    return new $errorClass($response['error']['message'], $data['error']['context'] ?? [], $e);
+                }
+            }
+
+            return null;
+        } catch (\Throwable $decodingError) {
+            return null;
+        }
     }
 
     /**